requirements.txt
boto3==1.34.14
paramiko==3.4.0
mysql-connector-python==8.3.0
Flask==3.0.0
requests==2.31.0
PyJWT==2.8.0
cryptography==41.0.7